Organize syncing classes into sync package.
