+++ app/src/main/java/com/bsstokes/bsdiy/MainActivity.java
@@ -19,6 +19,7 @@ import com.bsstokes.bsdiy.explore.ExploreFragment;
 import com.bsstokes.bsdiy.messages.MessagesFragment;
 import com.bsstokes.bsdiy.skills.SkillsFragment;
 import com.bsstokes.bsdiy.stream.StreamFragment;
+import com.bsstokes.bsdiy.sync.ApiSyncService;
 import com.bsstokes.bsdiy.to_dos.ToDosFragment;
 
 import butterknife.BindView;
