+++ app/src/main/java/com/bsstokes/bsdiy/sync/ApiSyncService.java
@@ -1,4 +1,4 @@
-package com.bsstokes.bsdiy;
+package com.bsstokes.bsdiy.sync;
 
 import android.app.IntentService;
 import android.content.Context;
